backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
tqdm
skyfield
openai
python-dotenv

